requirements.txt
aiogram==3.0.0
SQLAlchemy==2.0.36
python-dotenv==1.0.0
requests==2.31.0
psycopg2==2.9.10
aiosqlite==0.19.0  
python-telegram-bot==20.5  